aaaaaaaaaaa

+++ app/Services/Admin/Moderator/Service.php
@@ -6,9 +6,21 @@ namespace App\Services\Admin\Moderator;
 use App\Models\Artist;
 use App\Models\ArtistAudit;
 use App\Models\Track;
+use App\Services\Admin\HelperService;
+use OwenIt\Auditing\Models\Audit;
+use PhpParser\Node\Stmt\Break_;
+use PhpParser\Node\Stmt\Continue_;
 
 class Service
 {
+
+    private $helper;
+
+    public function __construct()
+    {
+        $this->helper = new HelperService();
+    }
+
     public function create($audit)
     {
         if (preg_match('/(Artist)/', $audit->auditable_type))
@@ -72,7 +84,7 @@ class Service
             $path = substr($path, 0, strpos($path, basename($path)));
             $path = pathToServer() . substr($path, strpos($path, "images"));
             $path = preg_replace('/artist_artWork\//', '', $path);
-            $this->deletePath($path);
+            $this->helper->delete($path);
         }
 
         if (isset($artist))
@@ -84,26 +96,32 @@ class Service
 
     private function deleteTrack($audit)
     {
-
-        if ($audit->event == 'deleted')
+        if ($audit->event == 'deleted') {
             $track =  Track::withTrashed()->where('id',  $audit->old_values['id'])->get()[0];
-        else
+        }
+        if ($audit->event == 'created') {
             $track = Track::where('id', $audit->new_values['id'])->get()[0];
+        }
+        if ($audit->event == 'updated')
+            $artwork_url = $audit->old_values['artwork_url'];
 
-        if (($path = $track->artwork_url)) {
+        if (($path = $track->artwork_url ?? $artwork_url)) {
             $path = substr($path, 0, strpos($path, basename($path)));
             $path = pathToServer() . substr($path, strpos($path, "images"));
             $path = preg_replace('/artwork\//', '', $path);
-            $this->deletePath($path);
+            $this->helper->delete($path);
         }
-        if (isset($track->albums))
-            $track->albums()->detach();
-        if (isset($track->artists));
-        $track->artists()->detach();
-        if (isset($track->genres))
-            $track->genres()->detach();
 
-        $track->delete();
+        if ($audit->event == 'deleted' || $audit->event == 'created') {
+            if (isset($track->albums))
+                $track->albums()->detach();
+            if (isset($track->artists));
+            $track->artists()->detach();
+            if (isset($track->genres))
+                $track->genres()->detach();
+
+            $track->delete();
+        }
     }
 
 
@@ -132,27 +150,78 @@ class Service
     private function dataTrack($audit)
     {
         $data = $audit->old_values;
+
+        preg_match('/\d+$/', $audit->url, $id);
+
         $data['status'] = true;
 
-        $track = Track::where('title', $audit->new_values['title'])->get();
-        $track[0]->update($data);
+        $track = Track::where('id', $id[0])->get()[0];
+
+        if ($audit->old_values != $audit->new_values)
+            $this->helper->delete($track->artwork_url);
+
+        $track->update($data);
     }
 
-    private function deletePath($path)
+    public function deleteEvents($audits)
     {
+        // preg_match('/\d+$/', Audit::latest()->first()->url, $id);
 
-        if (is_dir($path) === true) {
-            $files = array_diff(scandir($path), array('.', '..'));
+        // Проеряем на совпадение на моделей
+        while ($audits = $this->similarAudit($audits)) {
+            // последние изменение трека
 
-            foreach ($files as $file) {
-                $this->deletePath(realpath($path) . '/' . $file);
+            // if (count($audits) === 1) continue;
+
+            $recentСhanges = $audits->latest()->first();
+            // все изменние трека кроме послденего
+            $allChanges = $audits->latest()->get()->pop();
+
+            // если нет измение возращаем false
+            if (count($allChanges) < 0) return false;
+
+            // если есть измение кроме последнего отменяем действие
+            foreach ($allChanges as $audit) {
+                // если это удаленный трек востанавливаем \
+                if ($audit->event == 'deleted')
+                    $this->create($audit);
+                // если трек изменен возращаем прежние данные  \
+                if ($audit->event == 'updated')
+                    $this->returnOldData($audit);
             }
-
-            return rmdir($path);
-        } else if (is_file($path) === true) {
-            return unlink($path);
         }
 
-        return false;
+        dd($audits);
+        return $audits;
+    }
+
+
+    /**
+     * Находит сибытие на одинаковые модели
+     *
+     * @param object  $audits
+     * @return array || null
+     */
+    private function similarAudit(&$audits)
+    {
+        $firstId = null;
+        $secondId = null;
+        $similar = [];
+
+        foreach ($audits as $audit)
+            if (preg_match('/(\d+)$/', $audit->url, $firstId))
+                break;
+
+        foreach ($audits as $audit) {
+            if (preg_match('/(\d+)$/', $audit->url, $secondId)) {
+
+                if ($firstId[0] === $secondId[0]) {
+                    $similar[] = $audit;
+                    unset($audit);
+                }
+            }
+        }
+
+        return $similar;
     }
 }
